# Conflicts:
#	limes-core/pom.xml
#	limes-core/src/main/java/org/aksw/limes/core/io/ls/LinkSpecification.java
#	limes-core/src/main/java/org/aksw/limes/core/measures/mapper/topology/RADON.java
#	limes-core/src/main/java/org/aksw/limes/core/measures/measure/MeasureFactory.java
#	limes-core/src/main/java/org/aksw/limes/core/ml/algorithm/Eagle.java
#	limes-core/src/main/java/org/aksw/limes/core/ml/algorithm/euclid/LinearEuclid.java

+++ limes-core/src/main/java/org/aksw/limes/core/measures/measure/space/AMinkowskiMeasure.java
@@ -1,3 +1,20 @@
+/*
+ * LIMES Core Library - LIMES – Link Discovery Framework for Metric Spaces.
+ * Copyright © 2011 Data Science Group (DICE) (ngonga@uni-paderborn.de)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package org.aksw.limes.core.measures.measure.space;
 
 import org.aksw.limes.core.io.cache.Instance;
@@ -44,7 +61,7 @@ abstract class AMinkowskiMeasure extends ASpaceMeasure {
 
         double distance = 0;
         for (int i = 0; i < p1.length; i++) {
-            double min = Integer.MAX_VALUE;
+            double min = Double.MAX_VALUE;
             for (String value1 : instance1.getProperty(p1[i])) {
                 for (String value2 : instance2.getProperty(p2[i])) {
                     try {
@@ -55,7 +72,7 @@ abstract class AMinkowskiMeasure extends ASpaceMeasure {
             }
             distance = distance + min;
         }
-        return 1.0 / (1 + outerPTerm(distance));
+        return 1.0 / (1.0 + outerPTerm(distance));
     }
 
     /**
